+++ session/commands.go
@@ -2,7 +2,7 @@ package session
 
 import (
 	"bytes"
-	"fmt"
+	"errors"
 
 	"github.com/jedib0t/go-pretty/v6/table"
 )
@@ -22,23 +22,22 @@ type commandSet struct {
 	l []*command
 }
 
-func (c *commandSet) register(cmd_string string, cmd *command) {
+func (c *commandSet) register(cmd_string string, cmd *command) error {
 	_, ok := c.m[cmd_string]
 	if ok {
-		fmt.Printf("warning: command %v has already been defined!\n", cmd_string)
-		return
+		return errors.New("command " + cmd_string + " has already been defined!\n")
 	}
 
 	c.m[cmd_string] = cmd
 
 	for _, command := range c.l {
 		if command == cmd {
-			return
+			return nil
 		}
 	}
 
 	c.l = append(c.l, cmd)
-
+	return nil
 }
 
 // func (c commandSet) has_command(cmd_string string) bool {
@@ -135,7 +134,7 @@ func newCommandSet() *commandSet {
 
 var commands *commandSet
 
-func (s *Session) init() { // to avoid cycle
+func (s *Session) init_commands() error {
 
 	commands = newCommandSet()
 
@@ -153,8 +152,12 @@ func (s *Session) init() { // to avoid cycle
 		},
 	}
 
-	commands.register("help", c_help)
-	commands.register("h", c_help)
+	if err := commands.register("help", c_help); err != nil {
+		return err
+	}
+	if err := commands.register("h", c_help); err != nil {
+		return err
+	}
 
 	//quit
 	c_quit := &command{
@@ -168,14 +171,19 @@ func (s *Session) init() { // to avoid cycle
 		},
 	}
 
-	commands.register("quit", c_quit)
-	commands.register("q", c_quit)
+	if err := commands.register("quit", c_quit); err != nil {
+		return err
+	}
+	if err := commands.register("q", c_quit); err != nil {
+		return err
+	}
 
 	// clearscreen
 
+	cl_func := s.exec_clearscreen()
 	c_clearscreen := &command{
 		name:   "cl[earscreen]",
-		single: s.exec_clearscreen(),
+		single: cl_func,
 		usage: []struct {
 			args string
 			msg  string
@@ -183,9 +191,12 @@ func (s *Session) init() { // to avoid cycle
 			{"~", "clear the terminal screen"},
 		},
 	}
-
-	commands.register("clearscreen", c_clearscreen)
-	commands.register("cl", c_clearscreen)
+	if err := commands.register("clearscreen", c_clearscreen); err != nil {
+		return err
+	}
+	if err := commands.register("cl", c_clearscreen); err != nil {
+		return err
+	}
 
 	// environment: list
 	c_list := &command{
@@ -195,12 +206,15 @@ func (s *Session) init() { // to avoid cycle
 			args string
 			msg  string
 		}{
-			{"~", "list all identifiers in the environment alphabetically \n\t with types and values"},
+			{"~", "list all identifiers in the environment alphabetically\n\t with types and values"},
 		},
 	}
-
-	commands.register("list", c_list)
-	commands.register("l", c_list)
+	if err := commands.register("list", c_list); err != nil {
+		return err
+	}
+	if err := commands.register("l", c_list); err != nil {
+		return err
+	}
 
 	// environment: clear
 	c_clear := &command{
@@ -214,8 +228,12 @@ func (s *Session) init() { // to avoid cycle
 		},
 	}
 
-	commands.register("clear", c_clear)
-	commands.register("c", c_clear)
+	if err := commands.register("clear", c_clear); err != nil {
+		return err
+	}
+	if err := commands.register("c", c_clear); err != nil {
+		return err
+	}
 
 	// paste
 	c_paste := &command{
@@ -243,9 +261,12 @@ func (s *Session) init() { // to avoid cycle
 		},
 	}
 
-	commands.register("expression", c_expr)
-	commands.register("expr", c_expr)
-
+	if err := commands.register("expression", c_expr); err != nil {
+		return err
+	}
+	if err := commands.register("expr", c_expr); err != nil {
+		return err
+	}
 	// level: statement
 	c_stmt := &command{
 		name:     "stmt|statement",
@@ -257,9 +278,12 @@ func (s *Session) init() { // to avoid cycle
 			{"~ <input>", "expect <input> to be a statement"},
 		},
 	}
-
-	commands.register("statement", c_stmt)
-	commands.register("stmt", c_stmt)
+	if err := commands.register("statement", c_stmt); err != nil {
+		return err
+	}
+	if err := commands.register("stmt", c_stmt); err != nil {
+		return err
+	}
 
 	// level: program
 	c_prog := &command{
@@ -273,8 +297,12 @@ func (s *Session) init() { // to avoid cycle
 		},
 	}
 
-	commands.register("program", c_prog)
-	commands.register("prog", c_prog)
+	if err := commands.register("program", c_prog); err != nil {
+		return err
+	}
+	if err := commands.register("prog", c_prog); err != nil {
+		return err
+	}
 
 	// process: parse
 	c_parse := &command{
@@ -284,12 +312,16 @@ func (s *Session) init() { // to avoid cycle
 			args string
 			msg  string
 		}{
-			{"~ <input>", "parse <input>"},
+			{"~ <input>", "print string representation of ast <input> is parsed to\n\t --> Node-method: String() string"},
 		},
 	}
 
-	commands.register("parse", c_parse)
-	commands.register("p", c_parse)
+	if err := commands.register("parse", c_parse); err != nil {
+		return err
+	}
+	if err := commands.register("p", c_parse); err != nil {
+		return err
+	}
 
 	// process: parsetree: TODO
 	c_parsetree := &command{
@@ -299,12 +331,16 @@ func (s *Session) init() { // to avoid cycle
 			args string
 			msg  string
 		}{
-			{"~ <input>", "parse <input>"}, //TODO
+			{"~ <input>", "print tree representation of <input>' ast\n\t to all set displays "},
 		},
 	}
 
-	commands.register("parsetree", c_parsetree)
-	commands.register("ptree", c_parsetree)
+	if err := commands.register("parsetree", c_parsetree); err != nil {
+		return err
+	}
+	if err := commands.register("ptree", c_parsetree); err != nil {
+		return err
+	}
 
 	// process: eval
 	c_eval := &command{
@@ -314,12 +350,15 @@ func (s *Session) init() { // to avoid cycle
 			args string
 			msg  string
 		}{
-			{"~ <input>", "print out value of object <input> evaluates to"},
+			{"~ <input>", "print value of object <input> evaluates to\n\t --> Object-method: Inspect() string"},
 		},
 	}
-
-	commands.register("eval", c_eval)
-	commands.register("e", c_eval)
+	if err := commands.register("eval", c_eval); err != nil {
+		return err
+	}
+	if err := commands.register("e", c_eval); err != nil {
+		return err
+	}
 
 	// process: type
 	c_type := &command{
@@ -332,24 +371,31 @@ func (s *Session) init() { // to avoid cycle
 			{"~ <input>", "show objecttype <input> evaluates to"},
 		},
 	}
-
-	commands.register("type", c_type)
-	commands.register("t", c_type)
+	if err := commands.register("type", c_type); err != nil {
+		return err
+	}
+	if err := commands.register("t", c_type); err != nil {
+		return err
+	}
 
 	// process: trace
 	c_trace := &command{
-		name:     "trace",
+		name:     "tr[ace]",
 		with_arg: s.exec_trace,
 		usage: []struct {
 			args string
 			msg  string
 		}{
-			{"~ <input>", "show evaluation trace step by step"},
+			{"~ <input>", "show evaluation trace interactively step by step"},
 		},
 	}
-	commands.register("trace", c_trace)
-
-	// process: evaltree TODO
+	if err := commands.register("trace", c_trace); err != nil {
+		return err
+	}
+	if err := commands.register("tr", c_trace); err != nil {
+		return err
+	}
+	// process: evaltree
 	c_evaltree := &command{
 		name:     "e[val]tree",
 		with_arg: s.exec_evaltree,
@@ -357,12 +403,15 @@ func (s *Session) init() { // to avoid cycle
 			args string
 			msg  string
 		}{
-			{"~ <input>", "print out value of object <input> evaluates to"}, //TODO
+			{"~ <input>", "print annotated tree representation of <input>'s ast\n\t to all set displays "},
 		},
 	}
-
-	commands.register("evaltree", c_evaltree)
-	commands.register("etree", c_evaltree)
+	if err := commands.register("evaltree", c_evaltree); err != nil {
+		return err
+	}
+	if err := commands.register("etree", c_evaltree); err != nil {
+		return err
+	}
 
 	// settings: settings
 	c_settings := &command{
@@ -372,10 +421,12 @@ func (s *Session) init() { // to avoid cycle
 			args string
 			msg  string
 		}{
-			{"~", "list all settings with their current values and defaults"},
+			{"~", "list all settings with their current and default values"},
 		},
 	}
-	commands.register("settings", c_settings)
+	if err := commands.register("settings", c_settings); err != nil {
+		return err
+	}
 
 	// settings: set
 	c_set := &command{
@@ -385,32 +436,38 @@ func (s *Session) init() { // to avoid cycle
 			args string
 			msg  string
 		}{
-			{"~ process <p>", "<p> must be: [e]val, [p]arse, [t]ype"},
-			{"~ level <l>", "<l> must be: [p]rogram, [s]tatement, [e]xpression"},
-			{"~ logparse", "additionally output ast-string"},
-			{"~ logtype", "additionally output objecttype"},
-			{"~ logtrace", "additionally output evaluation trace"},
-			{"~ incltoken", "include tokens in representations of asts"},
-			{"~ paste", "enable multiline support"},
 			{"~ prompt <prompt>", "set prompt string to <prompt>"},
-			{"~ treefile <f>", "set file that outputs ast-pdfs to <f>"},
-			{"~ evalfile <f>", "set file that outputs eval-pdfs to <f>"},
+			{"~ paste", "enable multiline support"},
+			{"~ level <l>", "<l> must be: p[rogram], s[tatement], e[xpression]"},
+			{"~ process <p>", "<p> must be: p[arse], p[arse]tree, e[val], e[val]tree,\n\t [t]ype, [tr]ace"},
+			{"~ logs <+|-l_0...+|-l_n>", "<l_i> must be: p[arse]tree, e[val]tree, [t]ype, [tr]ace"},
+			{"~ displays <+|-l_0...+|-l_n>", "<l_i> must be: p[arse]tree, e[val]tree, [t]ype, [tr]ace"},
+			{"~ verbosity <v>", "<v> must be 0, 1, 2"},
+			{"~ inclToken", "include tokens in representations of asts"},
+			{"~ inclEnv", "include environments in representations of asts"},
+			{"~ file <f>", "set file for pdfs to <f>"},
 		},
 	}
-	commands.register("set", c_set)
+	if err := commands.register("set", c_set); err != nil {
+		return err
+	}
 
 	// settings: reset
 	c_reset := &command{
 		name:     "reset",
+		single:   s.exec_reset_all,
 		with_arg: s.exec_reset,
 		usage: []struct {
 			args string
 			msg  string
 		}{
-			{"~ <setting>", "set <setting> to default\n\t for an overview consult :settings and/or :h set"},
+			{"~", "reset all settings"},
+			{"~ <setting>", "set <setting> to default value\n\t for an overview consult :settings and/or :h set"},
 		},
 	}
-	commands.register("reset", c_reset)
+	if err := commands.register("reset", c_reset); err != nil {
+		return err
+	}
 
 	// settings: unset
 	c_unset := &command{
@@ -427,5 +484,8 @@ func (s *Session) init() { // to avoid cycle
 			//incltoken logtrace
 		},
 	}
-	commands.register("unset", c_unset)
+	if err := commands.register("unset", c_unset); err != nil {
+		return err
+	}
+	return nil
 }